requirements.txt
# Google Cloud dependencies (API-Only Mode)
google-auth>=2.16.0
google-api-python-client>=2.100.0

# Database
psycopg2-binary>=2.9.6
sqlalchemy>=2.0.20
asyncpg>=0.28.0

# Utilities
python-dotenv>=1.0.0
pandas>=2.0.3
pyarrow>=12.0.1
pyyaml>=6.0.0
aiofiles>=23.0.0

# API and scheduling
requests>=2.31.0
APScheduler>=3.10.1

# MCP server dependencies
fastapi>=0.100.0
uvicorn>=0.23.0
pydantic>=2.0.0

# Metrics and observability
flask>=3.0.0
gunicorn>=21.2.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0
pytest-asyncio>=0.21.0
